requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.20
playwright>=1.40
playwright-stealth>=1.0
openpyxl>=3.1
pyyaml>=6.0
pydantic>=2.0
